slight refactor in order to support dropping column families
